Update Board.java
Fixes: adjusted collides to align with the well and panel. in particular, added instead of subtracted.

adjusted deleteRow : first, clear the well of a row which has been established as full - then, through a nested for loop, take each row above and move it down in the well. 

deleteRows - fixed a casting issue and now calls deleteRow on the relevant row(s)

getCompletedRows - fixed another casting issue - .add was pulling from well instead of the integer row, so the linked list was not functioning properly somehow? i'm not really sure as to why